requirements.txt
tensorflow==2.17.0
flask==3.0.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
werkzeug==3.0.1
gunicorn==23.0.0
h5py==3.11.0